refactor(api): restructure user module, update todo model with new fields, and refactor auth services
